pyproject.toml
[project]
name = "global-weather-forecasting"
version = "0.1.0"
description = "End-to-end weather trend forecasting using EDA, anomaly detection, time series & ensemble ML models on the Kaggle Global Weather Repository dataset."
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "xgboost>=1.7.0",
    "plotly>=5.15.0",
    "kaggle>=1.5.0",
    "joblib>=1.3.0",
    "gradio>=4.0.0",
    "prophet>=1.1.0",
    "scipy>=1.11.0",
    "lightgbm>=4.0.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "modal>=0.73.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
